+++ tests/read_test.rs
@@ -5,8 +5,8 @@ use exif::parse_exif;
 use failure;
 
 use libheif_rs::{
-    check_file_type, Chroma, ColorSpace, CompressionFormat, EncoderParameterValue, EncoderQuality,
-    FileTypeResult, HeifContext, StreamReader,
+    check_file_type, Channel, Chroma, ColorSpace, CompressionFormat, EncoderParameterValue,
+    EncoderQuality, FileTypeResult, HeifContext, RgbChroma, StreamReader,
 };
 
 #[test]
@@ -31,9 +31,8 @@ fn read_from_reader() -> Result<(), failure::Error> {
     assert_eq!(handle.width(), 3024);
     assert_eq!(handle.height(), 4032);
 
-    let src_img = handle.decode(ColorSpace::Undefined, Chroma::Undefined)?;
-    assert_eq!(src_img.color_space(), ColorSpace::YCbCr);
-    assert_eq!(src_img.chroma_format(), Chroma::C420);
+    let src_img = handle.decode(ColorSpace::Undefined)?;
+    assert_eq!(src_img.color_space(), Some(ColorSpace::YCbCr(Chroma::C420)));
 
     Ok(())
 }
@@ -114,13 +113,20 @@ fn decode_and_scale_image() -> Result<(), failure::Error> {
     let handle = ctx.primary_image_handle()?;
 
     // Decode the image
-    let src_img = handle.decode(ColorSpace::Undefined, Chroma::Undefined)?;
-    assert_eq!(src_img.color_space(), ColorSpace::YCbCr);
-    assert_eq!(src_img.chroma_format(), Chroma::C420);
+    let src_img = handle.decode(ColorSpace::Undefined)?;
+    assert_eq!(src_img.color_space(), Some(ColorSpace::YCbCr(Chroma::C420)));
     let planes = src_img.planes();
     let y_plane = planes.y.unwrap();
     assert_eq!(y_plane.width, 3024);
     assert_eq!(y_plane.height, 4032);
+
+    let cb_plane = planes.cb.unwrap();
+    assert_eq!(cb_plane.width, 3024 / 2);
+    assert_eq!(cb_plane.height, 4032 / 2);
+
+    let cr_plane = planes.cr.unwrap();
+    assert_eq!(cr_plane.width, 3024 / 2);
+    assert_eq!(cr_plane.height, 4032 / 2);
 
     // Scale the image
     let img = src_img.scale(1024, 800, None)?;
@@ -130,6 +136,14 @@ fn decode_and_scale_image() -> Result<(), failure::Error> {
     assert_eq!(y_plane.height, 800);
     assert!(!y_plane.data.is_empty());
     assert!(y_plane.stride > 0);
+
+    let cb_plane = planes.cb.unwrap();
+    assert_eq!(cb_plane.width, 1024 / 2);
+    assert_eq!(cb_plane.height, 800 / 2);
+
+    let cr_plane = planes.cr.unwrap();
+    assert_eq!(cr_plane.width, 1024 / 2);
+    assert_eq!(cr_plane.height, 800 / 2);
 
     Ok(())
 }
@@ -193,6 +207,40 @@ fn test_check_file_type() -> Result<(), failure::Error> {
     assert_eq!(check_file_type(&data[..7]), FileTypeResult::MayBe);
     assert_eq!(check_file_type(&data[..8]), FileTypeResult::MayBe);
     assert_eq!(check_file_type(&data[..12]), FileTypeResult::Supported);
+
+    Ok(())
+}
+
+#[test]
+fn read_test_from_readme_file() -> Result<(), failure::Error> {
+    let ctx = HeifContext::read_from_file("./data/test.heic")?;
+    let handle = ctx.primary_image_handle()?;
+    assert_eq!(handle.width(), 3024);
+    assert_eq!(handle.height(), 4032);
+
+    // Get Exif
+    let meta_ids = handle.list_of_metadata_block_ids("Exif", 1);
+    assert_eq!(meta_ids.len(), 1);
+    let exif: Vec<u8> = handle.metadata(meta_ids[0])?;
+
+    // Decode the image
+    let image = handle.decode(ColorSpace::Rgb(RgbChroma::Rgb))?;
+    assert_eq!(image.color_space(), Some(ColorSpace::Rgb(RgbChroma::Rgb)));
+    assert_eq!(image.width(Channel::Interleaved)?, 3024);
+    assert_eq!(image.height(Channel::Interleaved)?, 4032);
+
+    // Scale the image
+    let small_img = image.scale(1024, 800, None)?;
+    assert_eq!(small_img.width(Channel::Interleaved)?, 1024);
+    assert_eq!(small_img.height(Channel::Interleaved)?, 800);
+
+    // Get "pixels"
+    let planes = small_img.planes();
+    let interleaved_plane = planes.interleaved.unwrap();
+    assert_eq!(interleaved_plane.width, 1024);
+    assert_eq!(interleaved_plane.height, 800);
+    assert!(!interleaved_plane.data.is_empty());
+    assert!(interleaved_plane.stride > 0);
 
     Ok(())
 }